Bind calendar setup with markers
This was the setup that I wanted to do all the way but somehow did not
manage to do correctly. First I needed a way to get the agile cadence
labels and numbers into the markers data frame. Then, I can easily bind
the two data frames. Initially I did it the other way around and that
didn't work correctly.