Add calendar year variable for better imputation
In the default config I have/had a situation where I dated a marker in
week 51 of the present calendar year but due to the left_join()
imputation of the missing agile variables this got put into PI xx.1
instead of xx.4.

This commit fixes the basic calculation but introduces a new thing that
is dealt with in the next commit.