Fill empty slots in agile calendar
Because the basic calender is laid out only for the current year, the
dissynchrony between the calendar and agile year that is by default at
Philips results in missing values because items in the last week(s) of
the current calendar year could already belong to the next agile year.
However, no data is present for the next year.

This commit fixes that. Another fix could be to generate three
sequential agile calendars with the present year in the middle. This
actually a safer way to ensure correct filling, but more cumbersome to
implement.