Fix bug in finding dates for recurring events
Basing indicating recurring events by using increment or iteration week
numbers, I wanted to join in the default dates for those weeks and
presumed that dplyr::join would autofill the dates independent of
whether I was using increment or iteration week numbers. It does that
but I do need to do this in two steps because otherwise join() does not
find matching rows and doesn't know what/how to find the date.